+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. The format 
 - Added support for source address filtering
 - Added rule to permananetly allow icmp echo replies from any source
 - Added command line switch to etables to allow icmp echos into router
+- Added check for maximum prefix tuples reached and current usage in etables -L
 
 ## [0.2.5] - 2023-02-21